Move method to ForgeHooks
